+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/scheduling/Region.scala
@@ -6,7 +6,6 @@ import edu.uci.ics.amber.engine.architecture.scheduling.config.ResourceConfig
 import org.jgrapht.graph.{DefaultEdge, DirectedAcyclicGraph}
 import org.jgrapht.traverse.TopologicalOrderIterator
 
-import java.net.URI
 import scala.jdk.CollectionConverters.IteratorHasAsScala
 
 case class RegionLink(fromRegionId: RegionIdentity, toRegionId: RegionIdentity)
