+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/scheduling/config/ResourceConfig.scala
@@ -3,7 +3,6 @@ package edu.uci.ics.amber.engine.architecture.scheduling.config
 import edu.uci.ics.amber.core.virtualidentity.PhysicalOpIdentity
 import edu.uci.ics.amber.core.workflow.{GlobalPortIdentity, PhysicalLink}
 
-import java.net.URI
 
 case class ResourceConfig(
     operatorConfigs: Map[PhysicalOpIdentity, OperatorConfig] = Map.empty,
