+++ core/amber/src/main/scala/edu/uci/ics/texera/web/resource/dashboard/hub/HubResource.scala
@@ -2,21 +2,8 @@ package edu.uci.ics.texera.web.resource.dashboard.hub
 
 import edu.uci.ics.texera.dao.SqlServer
 import edu.uci.ics.texera.dao.jooq.generated.Tables._
-import HubResource.{
-  fetchDashboardDatasetsByDids,
-  fetchDashboardWorkflowsByWids,
-  getUserLCCount,
-  isLikedHelper,
-  recordLikeActivity,
-  recordUserActivity,
-  userRequest,
-  validateEntityType
-}
-import edu.uci.ics.texera.web.resource.dashboard.user.workflow.WorkflowResource.{
-  DashboardWorkflow,
-  baseWorkflowSelect,
-  mapWorkflowEntries
-}
+import HubResource.{fetchDashboardDatasetsByDids, fetchDashboardWorkflowsByWids, getUserLCCount, isLikedHelper, recordLikeActivity, recordUserActivity, userRequest, validateEntityType}
+import edu.uci.ics.texera.web.resource.dashboard.user.workflow.WorkflowResource.{DashboardWorkflow, baseWorkflowSelect, mapWorkflowEntries}
 import org.jooq.impl.DSL
 
 import java.util
@@ -26,12 +13,12 @@ import javax.ws.rs._
 import javax.ws.rs.core.{Context, MediaType}
 import scala.jdk.CollectionConverters._
 import EntityTables._
+import edu.uci.ics.texera.dao.jooq.generated.tables.Dataset.DATASET
+import edu.uci.ics.texera.dao.jooq.generated.tables.DatasetUserAccess.DATASET_USER_ACCESS
+import edu.uci.ics.texera.dao.jooq.generated.tables.User.USER
+import edu.uci.ics.texera.dao.jooq.generated.tables.pojos.{Dataset, DatasetUserAccess}
 import edu.uci.ics.texera.web.resource.dashboard.DashboardResource.DashboardClickableFileEntry
-import edu.uci.ics.texera.web.resource.dashboard.user.dataset.DatasetResource.{
-  DashboardDataset,
-  baseDatasetSelect,
-  mapDashboardDataset
-}
+import edu.uci.ics.texera.web.resource.dashboard.user.dataset.DatasetResource.DashboardDataset
 
 object HubResource {
   case class userRequest(entityId: Integer, userId: Integer, entityType: String)
@@ -262,7 +249,15 @@ object HubResource {
       return List.empty[DashboardDataset]
     }
 
-    val records = baseDatasetSelect()
+    val records = context
+      .select()
+      .from(
+        DATASET
+          .leftJoin(DATASET_USER_ACCESS)
+          .on(DATASET_USER_ACCESS.DID.eq(DATASET.DID))
+          .leftJoin(USER)
+          .on(USER.UID.eq(DATASET.OWNER_UID))
+      )
       .where(DATASET.DID.in(dids: _*))
       .groupBy(
         DATASET.DID,
@@ -276,7 +271,17 @@ object HubResource {
       )
       .fetch()
 
-    mapDashboardDataset(records, uid)
+    records.asScala.map { record =>
+      val dataset = record.into(DATASET).into(classOf[Dataset])
+      val datasetAccess = record.into(DATASET_USER_ACCESS).into(classOf[DatasetUserAccess])
+      val ownerEmail = record.into(USER).getEmail
+      DashboardDataset(
+        isOwner = if (uid == null) false else dataset.getOwnerUid == uid,
+        dataset = dataset,
+        accessPrivilege = datasetAccess.getPrivilege,
+        ownerEmail = ownerEmail,
+      )
+    }.toList
   }
 }
