+++ core/gui/src/app/dashboard/component/user/user-dataset/user-dataset-explorer/user-dataset-version-creator/user-dataset-version-creator.component.ts
@@ -145,7 +145,6 @@ export class UserDatasetVersionCreatorComponent implements OnInit {
         ownerUid: undefined,
         storagePath: undefined,
         creationTime: undefined,
-        versionHierarchy: undefined,
       };
       this.datasetService
         .createDataset(ds)
