+++ core/gui/src/app/dashboard/type/dashboard-dataset.interface.ts
@@ -6,9 +6,4 @@ export interface DashboardDataset {
   ownerEmail: string;
   dataset: Dataset;
   accessPrivilege: "READ" | "WRITE" | "NONE";
-  versions: {
-    datasetVersion: DatasetVersion;
-    fileNodes: DatasetFileNode[];
-  }[];
-  size?: number;
 }
